feat: msw 패키지 제거 후 api 패키지에 통합 (타입 관련 역할의 애매함)
