+++ apps/attendance/pages/_app.tsx
@@ -39,7 +39,7 @@ function App(props: AppProps) {
   );
   useEffect(() => {
     async function init() {
-      const { initMocks } = await import('@weekly/msw');
+      const { initMocks } = await import('@weekly/api');
       await initMocks();
       setShouldRender(true);
     }
