+++ apps/qr/next.config.js
@@ -3,12 +3,7 @@
 module.exports = {
   reactStrictMode: true,
   swcMinify: true,
-  transpilePackages: [
-    '@weekly/ui',
-    '@weekly/api',
-    '@weekly/utils',
-    '@weekly/msw',
-  ],
+  transpilePackages: ['@weekly/ui', '@weekly/api', '@weekly/utils'],
   i18n: {
     locales: ['ko'],
     defaultLocale: 'ko',
