+++ apps/qr/package.json
@@ -25,7 +25,6 @@
     "@types/node": "^17.0.12",
     "@types/react": "^18.0.22",
     "@types/react-dom": "^18.0.7",
-    "@weekly/msw": "*",
     "@weekly/tsconfig": "*",
     "eslint": "7.32.0",
     "eslint-config-weekly": "*",
