+++ packages/api/index.tsx
@@ -1,6 +1,14 @@
-export * from './lib/attendance';
-export * from './lib/auth';
-export * from './lib/error';
-export * from './lib/members';
-export * from './lib/QueryClientProvider';
-export * from './lib/sessions';
+async function initMocks() {
+  if (typeof window === 'undefined') {
+    const { server } = await import('./lib/msw/server');
+    server.listen();
+  } else {
+    const { worker } = await import('./lib/msw/browser');
+    worker.start();
+  }
+}
+
+initMocks();
+
+export { initMocks };
+export * from './lib';
