+++ packages/api/lib/attendance/variables.ts
