+++ packages/api/lib/auth/variables.ts
