+++ packages/api/lib/common/error.ts
