+++ packages/api/package.json
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^4.24.4",
-    "@weekly/msw": "*",
-    "axios": "^1.3.2"
+    "axios": "^1.3.2",
+    "msw": "^1.0.0"
   }
 }
