--- packages/msw/.lintstagedrc
@@ -1,4 +1,0 @@
-{
-  "*.{js,ts,jsx,tsx}": ["yarn lint:fix"],
-  "*.{js,ts,jsx,tsx,json}": ["prettier --write"]
-}
