--- packages/msw/lib/members/index.ts
@@ -1,2 +1,0 @@
-export * from './handlers';
-export * from './url';
