--- packages/msw/lib/types.ts
@@ -1,4 +1,0 @@
-export * from './attendance/types';
-export * from './auth/types';
-export * from './members/types';
-export * from './sessions/types';
