--- packages/msw/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@weekly/msw",
-  "version": "0.0.0",
-  "main": "./index.ts",
-  "types": "./index.ts",
-  "license": "MIT",
-  "scripts": {
-    "lint": "TIMING=1 eslint \"**/*.{ts,tsx}\"",
-    "lint:fix": "TIMING=1 eslint \"**/*.{ts,tsx}\" --fix"
-  },
-  "devDependencies": {
-    "@types/node": "^18.13.0",
-    "@weekly/tsconfig": "*",
-    "eslint": "^7.32.0",
-    "eslint-config-weekly": "*",
-    "msw": "^1.0.0",
-    "react": "^18.2.0",
-    "typescript": "^4.9.4"
-  }
-}
