--- packages/msw/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@weekly/tsconfig/react-library.json",
-  "include": ["."],
-  "exclude": ["dist", "build", "node_modules"],
-  "compilerOptions": {
-    "lib": ["ES2015", "DOM"]
-  }
-}
